+++ common/genericError.js
@@ -1,0 +1,1 @@
+module.exports = res => res.status(500).json({error: 'We were unable to process your request.'});
